+++ CHANGELOG.md
@@ -19,3 +19,5 @@
 - Add async versions of `Result<T>.Match` and `Result<T>.Switch`.
 
 - Rename `Result<T>.ToType` to `To` and allow attempting to convert to non-derived types.
+
+- Add ToString extensions supporting `IFormattable`.